Avoid code bloat from float->string conversion
 * Convert to a `usize` before converting to a string to pass as an attribute to JS, since the code for printing a floating point number is really big apparently.

+++ engine/src/lib.rs
@@ -784,7 +784,7 @@ pub fn handle_key_down(key: &str, control_pressed: bool, shift_pressed: bool) {
             set_attr(
                 note_data.dom_id,
                 "x",
-                &(new_start_beat * BEAT_LENGTH_PX).to_string(),
+                &((new_start_beat * BEAT_LENGTH_PX) as usize).to_string(),
             );
 
             note_data.start_beat = new_start_beat;
